Implement lazy loading for AddEditLocationModal and extract StudioSections component
